requirements.txt
Flask-SQLAlchemy
flask-WTF
email_validator
